requirements.txt
matrix-nio>=0.10.0
Markdown>=3.1.1
PyYAML>=5.1.2
tomli>=1.1.0; python_version < "3.11"
Wand
python-magic
peewee
python-dateutil
wordcloud
